Removed an addionally created test file and merged my changes to the existing one
